pyproject.toml
[project]
name = "naay"
version = "2025.12.5-0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "rbroderi", email = "richard.broderick@gmail.com" }
]
requires-python = ">=3.13"
dependencies = []

[project.optional-dependencies]
dev = [
    "maturin>=1.10.2",
    "ruamel.yaml>=0.18",
    "pyyaml>=6",
    "rust-just>=1.43.1",
    "types-pyyaml>=6.0.12.20250915",
    "mypy>=1.19.0",
    "pytest>=8.3",
]
beartype = [
    "beartype>=0.22.8",
]

[build-system]
requires = ["maturin>=1.10,<2"]
build-backend = "maturin"

[tool.maturin]
manifest-path = "naay-py/Cargo.toml"
python-packages = ["naay", "_naay_pure"]
module-name = "_naay_native"
python-source = "src"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
# Enable all rules by default.
select = ["ALL"]

# Ignore specific rules.
ignore = [
    "TC006",
    "B024",
    "PIE790",
    "T201",
    "PYI013",
    "TC003",
    "PLC0414",
    "ERA001",
    "T203",
    "ANN401",
    "TC002",
    "TD002",
    "TD003",
    "FIX002",
    "FBT001",
    "FBT002",
    "D203",
    "D212",
    "S101",
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in docstrings.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "ARG002",
    "ARG001",
    "ARG005",
    "SLF001",
]
